space function

+++ searchdetail/server.py
@@ -20,49 +20,54 @@ def my_form():
 
 @app.route('/home', methods=['POST'])
 def my_form_post():
-    country = request.form['input1']
-    city = request.form['input2']
-    keyword = request.form['input3']
+    country_striped = request.form['input1']
+    city_striped = request.form['input2']
+    keyword_striped = request.form['input3']
 
-    if (os.stat('country.txt').st_size != 0):
-        with open('country.txt', 'a') as f:
-            f.write('')
-    if(os.stat('city.txt').st_size != 0):
-        with open('city.txt', 'a') as f:
-            f.write('')
-    if (os.stat('keyword.txt').st_size != 0):
-        with open('keyword.txt', 'a') as f:
-            f.write('')
-    new_country=''
-    new_city=''
-    new_keyword=''
+    country = country_striped.strip()
+    city = city_striped.strip()
+    keyword = keyword_striped.strip()
 
-    for b in country:
-        if(b=='\n'):
-            new_country+=''
-        else:
-            new_country+=b
+    if country != '' and city != '' and keyword != '':
+        if (os.stat(os.path.abspath(os.curdir)+'\country.txt').st_size != 0):
+            with open(os.path.abspath(os.curdir)+'\country.txt', 'a') as f:
+                f.write('\n')
+        if(os.stat(os.path.abspath(os.curdir)+'\city.txt').st_size != 0):
+            with open(os.path.abspath(os.curdir)+'\city.txt', 'a') as f:
+                f.write('\n')
+        if (os.stat(os.path.abspath(os.curdir)+'\keyword.txt').st_size != 0):
+            with open(os.path.abspath(os.curdir)+'\keyword.txt', 'a') as f:
+                f.write('\n')
+        new_country=''
+        new_city=''
+        new_keyword=''
 
-    for b in city:
-        if(b=='\n'):
-            new_city+=''
-        else:
-            new_city+=b
+        for b in country:
+            if(b=='\n'):
+                new_country+=''
+            else:
+                new_country+=b
 
-    for b in keyword:
-        if(b=='\n'):
-            new_keyword += ''
-        else:
-            new_keyword += b
+        for b in city:
+            if(b=='\n'):
+                new_city+=''
+            else:
+                new_city+=b
 
-    if request.method == 'POST':
-        if country!='' and city!='' and keyword!='':
-           with open('country.txt', 'a') as f:
-                f.write(str(new_country))
-           with open('city.txt', 'a') as f:
-               f.write(str(new_city))
-           with open('keyword.txt', 'a') as f:
-               f.write(str(new_keyword))
+        for b in keyword:
+            if(b=='\n'):
+                new_keyword += ''
+            else:
+                new_keyword += b
+
+        if request.method == 'POST':
+            if country!='' and city!='' and keyword!='':
+               with open(os.path.abspath(os.curdir)+'\country.txt', 'a') as f:
+                    f.write(str(new_country))
+               with open(os.path.abspath(os.curdir)+'\city.txt', 'a') as f:
+                   f.write(str(new_city))
+               with open(os.path.abspath(os.curdir)+'\keyword.txt', 'a') as f:
+                   f.write(str(new_keyword))
     # return render_template('home.html', find=find,near=near)
     return redirect('http://127.0.0.1:5000/home')
 
@@ -94,7 +99,7 @@ def my_form_post():
 
 @app.route("/view")
 def view_get():
-    con = sqlite3.connect("searchdetail.db")
+    con = sqlite3.connect(os.path.abspath(os.curdir)+"\searchdetail.db")
     con.row_factory = sqlite3.Row
     cur = con.cursor()
     cur.execute("select * from detail")
@@ -104,7 +109,7 @@ def view_get():
 
 @app.route("/delete")
 def delete_all():
-    con = sqlite3.connect("searchdetail.db")
+    con = sqlite3.connect(os.path.abspath(os.curdir)+"\searchdetail.db")
     con.row_factory = sqlite3.Row
     cur = con.cursor()
     cur.execute("select * from detail")
@@ -113,52 +118,59 @@ def delete_all():
     # rows = cur.fetchall()
     con.commit()
 
-    return render_template("home.html")
+    return redirect('http://127.0.0.1:5000/view')
 
 
 @app.route("/csv")
 def csv():
-    con = sqlite3.connect("searchdetail.db")
+    con = sqlite3.connect(os.path.abspath(os.curdir)+"\searchdetail.db")
     con.row_factory = sqlite3.Row
 
     df = pd.read_sql_query("SELECT * FROM detail", con)
     print(df)
     print(type(df))
-    df.to_csv('details.csv', index=False)
+    df.to_csv(os.path.abspath(os.curdir)+'\details.csv', index=False)
 
-    return redirect('http://127.0.0.1:5000/home')
+    return redirect('http://127.0.0.1:5000/view')
 
 
 @app.route("/deletebysearch", methods=['POST'])
 def deletebysearch():
-    key = request.form['del']
+    key_striped = request.form['del']
+
+    key = key_striped.strip()
     print(key)
     print(type(key))
-    sqliteConnection = sqlite3.connect('searchdetail.db')
-    cursor = sqliteConnection.cursor()
+    if(key!=''):
+        sqliteConnection = sqlite3.connect(os.path.abspath(os.curdir)+'\searchdetail.db')
+        cursor = sqliteConnection.cursor()
 
-    # sql_delete_query = "SELECT * FROM detail WHERE keyword LIKE '%Dentist%'"
-    # query = str("DELETE FROM detail WHERE keyword LIKE"+ " %{".format(key))
-    query = "DELETE FROM detail WHERE keyword LIKE '%{}%'".format(key)
-    sql_delete_query = query
-    cursor.execute(sql_delete_query)
-    sqliteConnection.commit()
+        # sql_delete_query = "SELECT * FROM detail WHERE keyword LIKE '%Dentist%'"
+        # query = str("DELETE FROM detail WHERE keyword LIKE"+ " %{".format(key))
+        query = "DELETE FROM detail WHERE keyword LIKE '%{}%'".format(key)
+        sql_delete_query = query
+        cursor.execute(sql_delete_query)
+        sqliteConnection.commit()
 
-    return render_template("home.html")
+    return redirect('http://127.0.0.1:5000/view')
 
 
 @app.route("/view", methods=['POST'])
 def view():
-    con = sqlite3.connect("searchdetail.db")
+    con = sqlite3.connect(os.path.abspath(os.curdir)+"\searchdetail.db")
     con.row_factory = sqlite3.Row
     cur = con.cursor()
 
 
 
     # if request.method == 'POST':
-    country = request.form['input1']
-    city = request.form['input2']
-    keyword = request.form['input3']
+    country_striped = request.form['input1']
+    city_striped = request.form['input2']
+    keyword_striped = request.form['input3']
+
+    country = country_striped.strip()
+    city = city_striped.strip()
+    keyword = keyword_striped.strip()
 
     if(country!='' and city!='' and keyword!=''):
         # cur.execute("select * from detail where country")
